+++ src/views/firebaseCMS/FirebaseCMS.tsx
@@ -1,26 +1,10 @@
 import React from "react";
 
-import { Authenticator, CMSApp, EntityCollection } from "@camberi/firecms";
+import { Authenticator, CMSApp } from "@camberi/firecms";
 import firebase from "firebase/app";
 import "typeface-rubik";
-
+import navigation from "./navigation";
 import "../../theme/FirebaseCMS.css";
-
-import userCollection from "./collections/user";
-import resourceCollection from "./collections/resource";
-import resourceTypesCollection from "./collections/resource_types";
-import officialCollection from "./collections/official";
-import partyCollection from "./collections/party";
-import cityCollection from "./collections/city";
-import stateCollection from "./collections/state";
-import rallyCollection from "./collections/rally";
-import topicCollection from "./collections/topic";
-import meetingCollection from "./collections/meeting";
-import meetingTypesCollection from "./collections/meeting_types";
-import stakeholderCollection from "./collections/stakeholder";
-import actionPlanCollection from "./collections/action_plan";
-import pageCollection from "./collections/page";
-import inviteCollection from "./collections/invite";
 
 // This is for test
 const firebaseConfig = {
@@ -33,24 +17,6 @@ const firebaseConfig = {
 };
 
 export function FirebaseCMS() {
-  const navigation: EntityCollection[] = [
-    userCollection,
-    resourceCollection,
-    resourceTypesCollection,
-    officialCollection,
-    partyCollection,
-    cityCollection,
-    stateCollection,
-    rallyCollection,
-    topicCollection,
-    meetingCollection,
-    meetingTypesCollection,
-    stakeholderCollection,
-    actionPlanCollection,
-    pageCollection,
-    inviteCollection
-  ];
-
   const myAuthenticator: Authenticator = (user?: firebase.User) => {
     console.log("Allowing access to", user?.email);
     return true;
